soc-simulator/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
redis[asyncio]==5.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
